Deprecated Functions in Package corpus

These functions are provided for compatibility with older versions of corpus only, and may be defunct as soon as the next release.

term_counts(x, filter = text_filter(x), weights = NULL,
            ngrams = NULL, min_count = NULL, max_count = NULL,
            min_support = NULL, max_support = NULL, types = FALSE)

Arguments

x
a text vector to tokenize.

filter
a token filter specifying the tokenization rules.

weights
a numeric vector the same length of x assigning weights to each text, or NULL for unit weights.

ngrams
an integer vector of n-gram lengths to include, or NULL for length-1 n-grams only.

min_count
a numeric scalar giving the minimum term count to include in the output, or NULL for no minimum count.

max_count
a numeric scalar giving the maximum term count to include in the output, or NULL for no maximum count.

min_support
a numeric scalar giving the minimum term support to include in the output, or NULL for no minimum support.

max_support
a numeric scalar giving the maximum term support to include in the output, or NULL for no maximum support.

types
a logical value indicating whether to include columns for the types that make up the terms.

Details

The term_counts function is deprecated, renamed to term_stats.

See also

Deprecated